git: add config
- Change git default branch name to main
- Add git git alias: when I type `git git` by mistake, this will make it work deduping it
- Configure git to sign all commits and tags with SSH key
- Show diff in commit message window
- Add `autocrlf = input`, so commited CRLF (supposedly by mistake) are automatically coverted to LF
- Display tabs with 4 spaces
- Set rebase.autosquash to true in git config. This is nice to have when committing with `--fixup` or `--squash` flags
- Set rerere.enabled: https://git-scm.com/book/en/v2/Git-Tools-Rerere

== Global gitignore

- Vim swap files
- Aactivator scripts
- Ctags
- .DS_Store
- Jetbrains `.idea`. dir

== Mark ignored revs in blame view

References:
https://www.michaelheap.com/git-ignore-rev/
https://git-scm.com/docs/git-blame#Documentation/git-blame.txt---ignore-revltrevgt

== Mergetool

Update to nvim, based on:
https://smittie.de/posts/git-mergetool/

== Ignore `.git` in global gitignore

This way nvim telescope doesn't show files in this folder.

I found in nvim docs:

	If you only want `file_ignore_patterns` for `find_files` and
	`grep_string`/`live_grep` it is suggested that you setup `gitignore`
	and have fd and or ripgrep installed because both tools will not show
	`gitignore`d files on default.

Git seems to omit `.git` but adding it to `.gitignore` still triggers
the behavior described above. This way I don't have to complicate my
neovim config with an additional pattern.

Also, I wasn't sure how the pattern plays with the default configs and
pickers, so I chose this way which seems to be much safer, esp. when
having a global gitignore.