gpg: add config file with settings that should be secure
Generally, short key formats are not safe, because they are only 32 bit,
which is very short in 2022. Long format is not much better, as it's
twice in size.

The best approach is to use key fingerprint which is much longer.
Settings in this file configure to always show fingerprints by default.

Reference: https://security.stackexchange.com/a/84281/270046

Also:

- Add config for GPG agent to keep password for one week. This is not
  too short, but also not too long. Seems reasonable. And TBH I'm
  rebooting my machine much more often.