all-repos: add config
Configure a structure with path: 'ws/private/repos'.

This is the simpliest approach which would work with my current setup.
A bit manual to setup on a new machine, so maybe I'll try to automate it
in the future.

Config options for all-repos are in the main readme in their GitHub
repo:
https://github.com/asottile/all-repos

Also:

- Prefix clone function with a comma. Use '_have' to only define them if
  dependencies are present.

- Add a separate clone function which uses all-repos. Rename the old
  version to indicate that it's readonly, as otherwise all-repos would
  clone this repository as long as it's on GitHub.